+++ packages/helpers/README.md
@@ -1,0 +1,5 @@
+# `@aksara-ui/helpers`
+
+> Helper components for the Aksara design system.
+
+This package is part of the Aksara monorepo. Please read the README in [the Aksara repository](https://github.com/kata-ai/aksara-ui) for detailed documentation about this package.
