+++ packages/helpers/package.json
@@ -1,13 +1,14 @@
 {
-  "name": "@aksara-ui/react",
-  "version": "1.0.1-canary.31",
-  "description": "React implementation of Aksara Design System, from Kata.ai.",
+  "name": "@aksara-ui/helpers",
+  "version": "1.0.1-canary.1",
+  "description": "Helpers components for the Aksara design system.",
+  "sideEffects": false,
   "main": "dist/index.js",
-  "module": "dist/core.esm.js",
+  "module": "dist/icons.esm.js",
   "typings": "dist/index.d.ts",
   "aksara:src": "./src/index.ts",
-  "repository": "https://github.com/kata-ai/aksara-ui.git",
   "author": "Kata.ai <dev@kata.ai>",
+  "homepage": "https://github.com/kata-ai/aksara-ui#readme",
   "license": "Apache-2.0",
   "files": [
     "dist"
@@ -17,53 +18,29 @@
   },
   "scripts": {
     "build": "tsdx build --transpileOnly --tsconfig tsconfig.build.json",
-    "build-storybook": "build-storybook --quiet",
     "clean": "rimraf dist",
     "format": "prettier --write \"src/**/*.{js,jsx,ts,tsx}\"",
     "lint": "eslint \"src/**/*.{js,jsx,ts,tsx}\"",
     "prepublish": "yarn build",
-    "report-coverage": "codecov -f ./coverage/*.json",
     "start": "tsdx watch --tsconfig tsconfig.build.json --verbose",
-    "storybook": "start-storybook -p 9001",
-    "storybook:build": "build-storybook --quiet",
-    "test": "jest",
-    "test:coverage": "jest --coverage",
-    "type-check": "tsc --emitDeclarationOnly"
+    "validate": "yarn lint && yarn type-check",
+    "report-coverage": "codecov -f ./coverage/*.json"
   },
   "dependencies": {
     "@aksara-ui/core": "^1.0.1-canary.4",
-    "@aksara-ui/helpers": "^1.0.1-canary.1",
-    "@radix-ui/react-accordion": "^0.1.5",
-    "@radix-ui/react-checkbox": "^0.1.4",
-    "@radix-ui/react-dropdown-menu": "^0.1.1",
-    "@radix-ui/react-id": "^0.1.1",
-    "@radix-ui/react-popover": "^0.1.1",
+    "clsx": "^1.1.1",
     "@radix-ui/react-portal": "^0.1.1",
-    "@radix-ui/react-radio-group": "^0.1.1",
-    "@radix-ui/react-scroll-area": "^0.1.3",
-    "@radix-ui/react-tooltip": "^0.1.1",
-    "@radix-ui/react-visually-hidden": "^0.1.1",
-    "@reach/accordion": "^0.16.1",
+    "react-transition-group": "^4.4.2",
+    "polished": "^4.1.3",
     "@styled-system/css": "^5.1.5",
     "@styled-system/should-forward-prop": "^5.1.5",
     "@styled-system/theme-get": "^5.1.2",
     "@types/styled-system": "^5.1.13",
     "@types/styled-system__css": "^5.0.16",
     "@types/styled-system__theme-get": "^5.0.2",
-    "clsx": "^1.1.1",
-    "csstype": "^3.0.9",
-    "dayzed": "^3.2.2",
-    "deepmerge": "^4.2.2",
-    "downshift": "^6.1.7",
+    "@radix-ui/react-visually-hidden": "^0.1.1",
     "focus-trap": "^6.6.0",
-    "polished": "^4.1.3",
-    "react-dropzone": "^11.3.2",
-    "react-fast-compare": "^3.2.0",
-    "react-ranger": "^2.1.0",
-    "react-transition-group": "^4.4.2",
-    "react-virtual": "^2.10.0",
-    "styled-system": "^5.1.4",
-    "tslib": "^2.3.1"
+    "styled-system": "^5.1.4"
   },
   "peerDependencies": {
     "@aksara-ui/icons": "^1.0.0-canary.0",
@@ -72,7 +49,6 @@
     "styled-components": "^5.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.0",
     "@storybook/addon-actions": "^6.3.7",
     "@storybook/addon-essentials": "^6.3.7",
     "@storybook/addon-knobs": "^6.3.0",
@@ -84,14 +60,9 @@
     "@types/jest": "^26.0.24",
     "@types/react-ranger": "^2.0.1",
     "@types/react-transition-group": "^4.4.4",
-    "babel-loader": "^8.2.2",
-    "babel-plugin-styled-components": "^1.13.2",
-    "babel-preset-react-app": "^10.0.0",
     "codecov": "^3.8.3",
     "copee": "^1.0.6",
     "jest": "^26.6.3",
-    "jest-styled-components": "^7.0.5",
-    "react-docgen-typescript-loader": "^3.6.0",
-    "react-table": "^7.7.0"
+    "jest-styled-components": "^7.0.5"
   }
 }
