+++ packages/helpers/src/overlay/index.ts
@@ -1,2 +1,3 @@
 // eslint-disable-next-line import/prefer-default-export
 export { default as Overlay } from './components/OverlayScreen';
+export * from './components/OverlayScreen';
