+++ packages/helpers/src/utils/useOnMount.ts
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export default function useOnMount(callback: React.EffectCallback) {
+  React.useEffect(callback, []);
+}
