+++ packages/react/src/components/navigation/components/PrimaryNavItem.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { VisuallyHidden } from '../../../helpers';
+import { VisuallyHidden } from '@aksara-ui/helpers';
 import { useComponentStyles } from '../../../system';
 import { UnstyledAnchor, UnstyledAnchorProps } from '../../../typography';
