+++ packages/react/src/components/slider/Slider.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { RangerOptions, useRanger } from 'react-ranger';
-import { VisuallyHidden } from '../../helpers';
+import { VisuallyHidden } from '@aksara-ui/helpers';
 import { Box, BoxProps } from '../../layout';
 import { useComponentStyles } from '../../system';
 import { UnstyledButton } from '../button';
