+++ packages/react/src/index.ts
@@ -1,6 +1,5 @@
 export * from './aksara-provider';
 export * from './components';
-export * from './helpers';
 export * from './layout';
 export * from './theme';
 export * from './typography';
